+++ website/docs/data-display/d3-chart/examples/chart-legend.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Line, minMax, XAxis, Plot, YAxis, ChartLegend } from 'intergalactic/d3-chart';
 import { scaleLinear } from 'd3-scale';
 import { Box } from 'intergalactic/flex-box';
-import Checkbox from 'intergalactic/checkbox';
 
 const Demo = () => {
   const MAP_THEME = {
@@ -63,6 +62,7 @@ const Demo = () => {
           onChangeVisibleItem={handleChangeVisible}
           onMouseEnterItem={handleMouseEnter}
           onMouseLeaveItem={handleMouseLeave}
+          aria-label={'Chart legend aria label'}
         />
       </Box>
       <Plot data={data} scale={[xScale, yScale]} width={width} height={height}>