+++ website/docs/data-display/d3-chart/examples/enforcing-patterns.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { Plot, XAxis, YAxis, minMax, StackedArea, HoverLine } from 'intergalactic/d3-chart';
+import { Plot, YAxis, minMax, StackedArea } from 'intergalactic/d3-chart';
 import { scaleLinear } from 'd3-scale';
-import { Flex, Box } from 'intergalactic/flex-box';
-import { Text } from 'intergalactic/typography';
 import { curveCardinal } from 'd3-shape';
 
 const Demo = () => {