+++ website/docs/data-display/line-chart/examples/basic-usage.tsx
@@ -9,6 +9,7 @@ const Demo = () => {
       plotHeight={200}
       groupKey={'x'}
       xTicksCount={data.length / 2}
+      aria-label={'Line chart'}
     />
   );
 };