+++ website/docs/data-display/radar-chart/examples/basic-usage.tsx
@@ -2,7 +2,15 @@ import React from 'react';
 import { Chart, colors } from 'intergalactic/d3-chart';
 
 const Demo = () => {
-  return <Chart.Radar data={data} groupKey={'categories'} plotWidth={400} plotHeight={400} />;
+  return (
+    <Chart.Radar
+      data={data}
+      groupKey={'categories'}
+      plotWidth={400}
+      plotHeight={400}
+      aria-label={'Radar chart'}
+    />
+  );
 };
 
 const data = {